test: move HMR main test to a separate file
